LayoutTests/imported/w3c/web-platform-tests/intersection-observer/initial-observation-with-threshold-expected.txt


PASS First observation with a threshold. 
FAIL First rAF assert_equals: entries[0].isIntersecting expected false but got true
PASS root.scrollTop = 20 
